Extract the hardcoded spec names into a `specs.py` file
